Manages the code coverage report in AJUR when the test storage service is available (i.e. when the `isTestServiceAvailable` instrumentation arg is set to be `true`). We leave to the testing infrastructure to decide whether the test service is available.
The coverage module (`InstrumentationCoverageReporter`) will be reused by the Flutter Instrumentation runner.

Design doc: go/axt-code-coverage

BEGIN_PUBLIC
Manages the code coverage report in AJUR when the test storage service is available (i.e. when the `isTestServiceAvailable` instrumentation arg is set to be `true`). We leave to the testing infrastructure to decide whether the test service is available.
END_PUBLIC

PiperOrigin-RevId: 343121991